# Conflicts:
#	dev/skeletons/skeleton_list.php
#	htdocs/accountancy/admin/export.php
#	htdocs/accountancy/bookkeeping/card.php
#	htdocs/adherents/card.php
#	htdocs/adherents/list.php
#	htdocs/adherents/type.php
#	htdocs/admin/dict.php
#	htdocs/admin/facture.php
#	htdocs/compta/bank/card.php
#	htdocs/compta/facture.php
#	htdocs/compta/facture/fiche-rec.php
#	htdocs/compta/facture/list.php
#	htdocs/compta/ventilation/card.php
#	htdocs/compta/ventilation/fournisseur/card.php
#	htdocs/contact/list.php
#	htdocs/core/class/html.formactions.class.php
#	htdocs/fichinter/list.php
#	htdocs/projet/card.php
#	htdocs/projet/list.php
#	htdocs/projet/tasks/list.php
#	htdocs/societe/notify/card.php
#	htdocs/societe/soc.php
#	htdocs/user/card.php
#	htdocs/user/index.php

+++ htdocs/accountancy/admin/fiscalyear_card.php
@@ -16,8 +16,9 @@
  */
 
 /**
- * \file htdocs/accountancy/admin/fiscalyear_card.php
- * \brief Page to show a fiscal year
+ * \file        htdocs/accountancy/admin/fiscalyear_card.php
+ * \ingroup     Advanced accountancy
+ * \brief       Page to show a fiscal year
  */
 require '../../main.inc.php';
 
@@ -28,7 +29,9 @@ $langs->load("admin");
 $langs->load("compta");
 
 // Security check
-if (! $user->admin)
+if ($user->societe_id > 0)
+	accessforbidden();
+if (empty($user->rights->accounting->fiscalyear))
 	accessforbidden();
 
 $error = 0;
@@ -71,13 +74,13 @@ if ($action == 'confirm_delete' && $confirm == "yes") {
 else if ($action == 'add') {
 	if (! GETPOST('cancel', 'alpha')) {
 		$error = 0;
-		
+
 		$object->date_start = $date_start;
 		$object->date_end = $date_end;
 		$object->label = GETPOST('label', 'alpha');
 		$object->statut = GETPOST('statut', 'int');
 		$object->datec = dol_now();
-		
+
 		if (empty($object->date_start) && empty($object->date_end)) {
 			setEventMessages($langs->trans("ErrorFieldRequired", $langs->transnoentitiesnoconv("Date")), null, 'errors');
 			$error ++;
@@ -86,20 +89,20 @@ else if ($action == 'add') {
 			setEventMessages($langs->trans("ErrorFieldRequired", $langs->transnoentitiesnoconv("Label")), null, 'errors');
 			$error ++;
 		}
-		
+
 		if (! $error) {
 			$db->begin();
-			
+
 			$id = $object->create($user);
-			
+
 			if ($id > 0) {
 				$db->commit();
-				
+
 				header("Location: " . $_SERVER["PHP_SELF"] . "?id=" . $id);
 				exit();
 			} else {
 				$db->rollback();
-				
+
 				setEventMessages($object->error, $object->errors, 'errors');
 				$action = 'create';
 			}
@@ -116,12 +119,12 @@ else if ($action == 'add') {
 else if ($action == 'update') {
 	if (! GETPOST('cancel', 'alpha')) {
 		$result = $object->fetch($id);
-		
+
 		$object->date_start = empty($_POST["fiscalyear"]) ? '' : $date_start;
 		$object->date_end = empty($_POST["fiscalyearend"]) ? '' : $date_end;
 		$object->label = GETPOST('label', 'alpha');
 		$object->statut = GETPOST('statut', 'int');
-		
+
 		$result = $object->update($user);
 		
 		if ($result > 0) {
@@ -136,108 +139,112 @@ else if ($action == 'update') {
 	}
 }
 
+
+
 /*
  * View
  */
 
-llxHeader();
+$title = $langs->trans("Fiscalyear") . " - " . $langs->trans("Card");
+$helpurl = "";
+llxHeader("",$title,$helpurl);
 
 $form = new Form($db);
 
 if ($action == 'create') 
 {
 	print load_fiche_titre($langs->trans("NewFiscalYear"));
-	
+
 	print '<form action="' . $_SERVER["PHP_SELF"] . '" method="POST">';
 	print '<input type="hidden" name="token" value="' . $_SESSION['newtoken'] . '">';
 	print '<input type="hidden" name="action" value="add">';
-	
+
 	dol_fiche_head();
-	
+
 	print '<table class="border" width="100%">';
-	
+
 	// Label
-	print '<tr><td class="fieldrequired">' . $langs->trans("Label") . '</td><td><input name="label" size="32" value="' . GETPOST("label") . '"></td></tr>';
-	
+	print '<tr><td class="titlefieldcreate fieldrequired">' . $langs->trans("Label") . '</td><td><input name="label" size="32" value="' . GETPOST("label") . '"></td></tr>';
+
 	// Date start
 	print '<tr><td class="fieldrequired">' . $langs->trans("DateStart") . '</td><td>';
 	print $form->select_date(($date_start ? $date_start : ''), 'fiscalyear');
 	print '</td></tr>';
-	
+
 	// Date end
 	print '<tr><td class="fieldrequired">' . $langs->trans("DateEnd") . '</td><td>';
 	print $form->select_date(($date_end ? $date_end : - 1), 'fiscalyearend');
 	print '</td></tr>';
-	
+
 	// Statut
 	print '<tr>';
 	print '<td class="fieldrequired">' . $langs->trans("Status") . '</td>';
 	print '<td class="valeur">';
 	print Form::selectarray('statut', $statut2label, GETPOST('statut'));
 	print '</td></tr>';
-	
+
 	print '</table>';
-	
+
 	dol_fiche_end();
-	
+
 	print '<div class="center">';
 	print '<input class="button" type="submit" value="' . $langs->trans("Save") . '">';
 	print '&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;';
 	print '<input class="button" type="submit" name="cancel" value="' . $langs->trans("Cancel") . '">';
 	print '</div>';
-	
+
 	print '</form>';
 } else if ($id) {
 	$result = $object->fetch($id);
 	if ($result > 0) {
 		$head = fiscalyear_prepare_head($object);
-		
+
 		if ($action == 'edit') {
-			dol_fiche_head($head, 'card', $langs->trans("FiscalYearCard"), 0, 'cron');
-			
+			dol_fiche_head($head, 'card', $langs->trans("Fiscalyear"), 0, 'cron');
+
 			print '<form name="update" action="' . $_SERVER["PHP_SELF"] . '" method="POST">' . "\n";
 			print '<input type="hidden" name="token" value="' . $_SESSION['newtoken'] . '">';
 			print '<input type="hidden" name="action" value="update">';
 			print '<input type="hidden" name="id" value="' . $id . '">';
-			
+
 			print '<table class="border" width="100%">';
-			
+
 			// Ref
 			print "<tr>";
-			print '<td class="titlefield">' . $langs->trans("Ref") . '</td><td>';
+			print '<td class="titlefieldcreate titlefield">' . $langs->trans("Ref") . '</td><td>';
 			print $object->ref;
 			print '</td></tr>';
-			
+
 			// Label
 			print '<tr><td class="fieldrequired">' . $langs->trans("Label") . '</td><td>';
 			print '<input name="label" class="flat" size="32" value="' . $object->label . '">';
 			print '</td></tr>';
-			
+
 			// Date start
 			print '<tr><td class="fieldrequired">' . $langs->trans("DateStart") . '</td><td>';
 			print $form->select_date($object->date_start ? $object->date_start : - 1, 'fiscalyear');
 			print '</td></tr>';
-			
+
 			// Date end
 			print '<tr><td class="fieldrequired">' . $langs->trans("DateEnd") . '</td><td>';
 			print $form->select_date($object->date_end ? $object->date_end : - 1, 'fiscalyearend');
 			print '</td></tr>';
-			
+
 			// Statut
 			print '<tr><td>' . $langs->trans("Statut") . '</td><td>';
 			print Form::selectarray('statut', $statut2label, $object->statut);
 			print '</td></tr>';
-			
+
 			print '</table>';
-			
+
 			print '<br><div class="center">';
 			print '<input type="submit" class="button" value="' . $langs->trans("Save") . '">';
 			print '&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;';
 			print '<input type="submit" name="cancel" class="button" value="' . $langs->trans("Cancel") . '">';
 			print '</div>';
-			
+
 			print '</form>';
-			
+
 			dol_fiche_end();
 		} else {
 			/*
@@ -246,59 +253,61 @@ if ($action == 'create')
 			if ($action == 'delete') {
 				print $form->formconfirm($_SERVER["PHP_SELF"] . "?id=" . $id, $langs->trans("DeleteFiscalYear"), $langs->trans("ConfirmDeleteFiscalYear"), "confirm_delete");
 			}
-			
-			dol_fiche_head($head, 'card', $langs->trans("FiscalYearCard"), 0, 'cron');
-			
+
+			dol_fiche_head($head, 'card', $langs->trans("Fiscalyear"), 0, 'cron');
+
 			print '<table class="border" width="100%">';
-			
+
 			$linkback = '<a href="' . DOL_URL_ROOT . '/accountancy/admin/fiscalyear.php">' . $langs->trans("BackToList") . '</a>';
-			
+
 			// Ref
-			print '<tr><td width="25%">' . $langs->trans("Ref") . '</td><td width="50%">';
+			print '<tr><td class="titlefield">' . $langs->trans("Ref") . '</td><td width="50%">';
 			print $object->ref;
-			print '</td><td width="25%">';
+			print '</td><td>';
 			print $linkback;
 			print '</td></tr>';
-			
+
 			// Label
 			print '<tr><td valign="top">';
 			print $form->editfieldkey("Label", 'label', $object->label, $object, $conf->global->MAIN_EDIT_ALSO_INLINE, 'alpha:32');
 			print '</td><td colspan="2">';
 			print $form->editfieldval("Label", 'label', $object->label, $object, $conf->global->MAIN_EDIT_ALSO_INLINE, 'alpha:32');
 			print "</td></tr>";
-			
+
 			// Date start
 			print '<tr><td>';
 			print $form->editfieldkey("DateStart", 'date_start', $object->date_start, $object, $conf->global->MAIN_EDIT_ALSO_INLINE, 'datepicker');
 			print '</td><td colspan="2">';
 			print $form->editfieldval("DateStart", 'date_start', $object->date_start, $object, $conf->global->MAIN_EDIT_ALSO_INLINE, 'datepicker');
 			print '</td></tr>';
-			
+
 			// Date end
 			print '<tr><td>';
 			print $form->editfieldkey("DateEnd", 'date_end', $object->date_end, $object, $conf->global->MAIN_EDIT_ALSO_INLINE, 'datepicker');
 			print '</td><td colspan="2">';
 			print $form->editfieldval("DateEnd", 'date_end', $object->date_end, $object, $conf->global->MAIN_EDIT_ALSO_INLINE, 'datepicker');
 			print '</td></tr>';
-			
+
 			// Statut
 			print '<tr><td>' . $langs->trans("Status") . '</td><td colspan="2">' . $object->getLibStatut(4) . '</td></tr>';
-			
+
 			print "</table>";
-			
+
 			dol_fiche_end();
+
+			if (! empty($user->rights->accounting->fiscalyear))
+			{
+    			/*
+    			 * Barre d'actions
+    			 */
+    			print '<div class="tabsAction">';
+
+    			print '<a class="butAction" href="' . $_SERVER["PHP_SELF"] . '?action=edit&id=' . $id . '">' . $langs->trans('Modify') . '</a>';
+    
+    			print '<a class="butActionDelete" href="' . $_SERVER["PHP_SELF"] . '?action=delete&id=' . $id . '">' . $langs->trans('Delete') . '</a>';
 			
-			/*
-			 * Barre d'actions
-			 */
-			
-			print '<div class="tabsAction">';
-			
-			print '<a class="butAction" href="' . $_SERVER["PHP_SELF"] . '?action=edit&id=' . $id . '">' . $langs->trans('Modify') . '</a>';
-			
-			print '<a class="butActionDelete" href="' . $_SERVER["PHP_SELF"] . '?action=delete&id=' . $id . '">' . $langs->trans('Delete') . '</a>';
-			
-			print '</div>';
+    			print '</div>';
+			}
 		}
 	} else {
 		dol_print_error($db);
